+++ packages/@vue/babel-preset-app/__tests__/babel-preset.spec.js
@@ -14,7 +14,7 @@ test('polyfill detection', () => {
       targets: { node: 'current' }
     }]]
   })
-  // default includes
+  // default i  ncludes
   expect(code).not.toMatch(`import "core-js/modules/es6.promise"`)
   // usage-based detection
   expect(code).not.toMatch(`import "core-js/modules/es6.map"`)
@@ -32,7 +32,7 @@ test('polyfill detection', () => {
   // promise polyfill alone doesn't work in IE, needs this as well. fix: #1642
   expect(code).toMatch(`import "core-js/modules/es6.array.iterator"`)
   // usage-based detection
-  expect(code).toMatch(`import "core-js/modules/es6.map"`)
+  expect(code).toMatch(/import _Map from ".*runtime-corejs2\/core-js\/map"/)
 })
 
 test('modern mode always skips polyfills', () => {
@@ -49,7 +49,7 @@ test('modern mode always skips polyfills', () => {
   // default includes
   expect(code).not.toMatch(`import "core-js/modules/es6.promise"`)
   // usage-based detection
-  expect(code).not.toMatch(`import "core-js/modules/es6.map"`)
+  expect(code).not.toMatch(/import _Map from ".*runtime-corejs2\/core-js\/map"/)
 
   ;({ code } = babel.transformSync(`
     const a = new Map()
@@ -63,7 +63,7 @@ test('modern mode always skips polyfills', () => {
   // default includes
   expect(code).not.toMatch(`import "core-js/modules/es6.promise"`)
   // usage-based detection
-  expect(code).not.toMatch(`import "core-js/modules/es6.map"`)
+  expect(code).not.toMatch(/import _Map from ".*runtime-corejs2\/core-js\/map"/)
   delete process.env.VUE_CLI_MODERN_BUILD
 })
 
@@ -92,7 +92,7 @@ test('async/await', () => {
   // should use regenerator runtime
   expect(code).toMatch(`import "regenerator-runtime/runtime"`)
   // should use required helper instead of inline
-  expect(code).toMatch(/@babel.*runtime\/helpers\/.*asyncToGenerator/)
+  expect(code).toMatch(/import _asyncToGenerator from ".*runtime-corejs2\/helpers\/esm\/asyncToGenerator\"/)
 })
 
 test('jsx', () => {
@@ -105,4 +105,36 @@ test('jsx', () => {
   `.trim(), defaultOptions)
   expect(code).toMatch(`var h = arguments[0]`)
   expect(code).toMatch(`return h("div", ["bar"])`)
+})
+
+test('jsx options', () => {
+  const { code } = babel.transformSync(`
+    export default {
+      render () {
+        return <div>bar</div>
+      }
+    }
+  `.trim(), {
+    babelrc: false,
+    presets: [[preset, {
+      jsx: {
+        injectH: false
+      }
+    }]]
+  })
+  expect(code).not.toMatch(`var h = arguments[0]`)
+  expect(code).toMatch(`return h("div", ["bar"])`)
+})
+
+test('disable absoluteRuntime', () => {
+  const { code } = babel.transformSync(`
+    const a = [...arr]
+  `.trim(), {
+    babelrc: false,
+    presets: [[preset, {
+      absoluteRuntime: false
+    }]]
+  })
+
+  expect(code).toMatch('import _toConsumableArray from "@babel/runtime-corejs2/helpers/esm/toConsumableArray"')
 })