+++ packages/@vue/cli-plugin-eslint/__tests__/eslintPlugin.spec.js
@@ -120,3 +120,23 @@ test('should not fix with --no-fix option', async () => {
   // files should not have been fixed
   expect(await read('src/main.js')).not.toMatch(';')
 })
+
+// #3167, #3243
+test('should not throw when src folder is ignored by .eslintignore', async () => {
+  const project = await create('eslint-ignore', {
+    plugins: {
+      '@vue/cli-plugin-babel': {},
+      '@vue/cli-plugin-eslint': {
+        config: 'airbnb',
+        lintOn: 'commit'
+      }
+    },
+    useConfigFiles: true
+  })
+
+  const { write, run } = project
+  await write('.eslintignore', 'src\n.eslintrc.js')
+
+  // should not throw
+  await run('vue-cli-service lint')
+})