+++ packages/@vue/cli-plugin-eslint/generator/index.js
@@ -20,7 +20,7 @@ module.exports = (api, { config, lintOn = [] }, _, invoking) => {
     devDependencies: {
       'babel-eslint': '^10.0.1',
       'eslint': '^5.8.0',
-      'eslint-plugin-vue': '^5.0.0-0'
+      'eslint-plugin-vue': '^5.0.0'
     }
   }
 
@@ -54,7 +54,7 @@ module.exports = (api, { config, lintOn = [] }, _, invoking) => {
   } else if (config === 'prettier') {
     eslintConfig.extends.push('@vue/prettier')
     Object.assign(pkg.devDependencies, {
-      '@vue/eslint-config-prettier': '^4.0.0'
+      '@vue/eslint-config-prettier': '^4.0.1'
     })
     // prettier & default config do not have any style rules
     // so no need to generate an editorconfig file
@@ -71,7 +71,7 @@ module.exports = (api, { config, lintOn = [] }, _, invoking) => {
 
   if (lintOn.includes('commit')) {
     Object.assign(pkg.devDependencies, {
-      'lint-staged': '^7.2.2'
+      'lint-staged': '^8.1.0'
     })
     pkg.gitHooks = {
       'pre-commit': 'lint-staged'
@@ -113,11 +113,11 @@ const applyTS = module.exports.applyTS = api => {
     eslintConfig: {
       extends: ['@vue/typescript'],
       parserOptions: {
-        parser: 'typescript-eslint-parser'
+        parser: '@typescript-eslint/parser'
       }
     },
     devDependencies: {
-      '@vue/eslint-config-typescript': '^3.2.0'
+      '@vue/eslint-config-typescript': '^4.0.0'
     }
   })
 }