+++ packages/@vue/cli-service/generator/index.js
@@ -9,10 +9,10 @@ module.exports = (api, options) => {
       'build': 'vue-cli-service build'
     },
     dependencies: {
-      'vue': '^2.5.17'
+      'vue': '^2.5.22'
     },
     devDependencies: {
-      'vue-template-compiler': '^2.5.17'
+      'vue-template-compiler': '^2.5.21'
     },
     'postcss': {
       'plugins': {
@@ -36,9 +36,19 @@ module.exports = (api, options) => {
 
   if (options.cssPreprocessor) {
     const deps = {
+      // TODO: remove 'sass' option in v4 or rename 'dart-sass' to 'sass'
       sass: {
         'node-sass': '^4.9.0',
-        'sass-loader': '^7.0.1'
+        'sass-loader': '^7.1.0'
+      },
+      'node-sass': {
+        'node-sass': '^4.9.0',
+        'sass-loader': '^7.1.0'
+      },
+      'dart-sass': {
+        fibers: '^3.1.1',
+        sass: '^1.16.0',
+        'sass-loader': '^7.1.0'
       },
       less: {
         'less': '^3.0.4',