+++ packages/@vue/cli-ui/src/util/folders.js
@@ -1,5 +1,5 @@
 export function isValidName (name) {
-  return !name.match(/[/@\s+%:]/) && encodeURIComponent(name) === name
+  return !name.match(/[/@\s+%:]|^[_.]/) && encodeURIComponent(name) === name && name.length <= 214
 }
 
 export function isValidMultiName (name) {